Imported from a different project directory
The purpose of this project is to learn backend development. The project is a to-do application connected to MongoDB, and it uses CRUD operations to modify task data based on their name and completion status, which is stored in JSON format. My plan is to rebuild the frontend and add a dark mode feature to enhance the user experience. Additionally, I would like to include a date field in the JSON data to improve the functionality of the application.

This project was created with the help of John Smilga and his video:https://www.youtube.com/watch?v=qwfE7fSVaZM&t=8576s